POLIO-2055: update WidgetPaper
- extract WidgetPaperRow for reuse
