+++ hat/assets/js/apps/Iaso/components/papers/WidgetPaperComponent.tsx
@@ -5,7 +5,15 @@ import React, {
     ReactNode,
 } from 'react';
 
-import { Paper, Divider, Typography, Grid, Collapse, Box } from '@mui/material';
+import {
+    Paper,
+    Divider,
+    Typography,
+    Grid,
+    Collapse,
+    Box,
+    SxProps,
+} from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import ExpandLess from '@mui/icons-material/ExpandLess';
 import ExpandMore from '@mui/icons-material/ExpandMore';
@@ -62,12 +70,14 @@ type Props = {
     className?: string;
     elevation?: number;
     children?: ReactNode;
+    sx?: SxProps;
 };
 export const WidgetPaper: FunctionComponent<Props> = ({
     IconButton,
     iconButtonProps,
     title,
     children,
+    sx,
     id = '',
     padded = false,
     expandable = false,
@@ -85,6 +95,7 @@ export const WidgetPaper: FunctionComponent<Props> = ({
     };
     return (
         <Paper
+            sx={sx}
             elevation={elevation}
             className={`${classes.root} ${className}`}
             id={id}
